+++ packages/theme/src/lib/legacy-theme.ts
@@ -508,6 +508,54 @@ export const legacyTheme = {
         maxHeight: '100px',
       },
     },
+    MuiCheckbox: {
+      defaultProps: {
+        disableRipple: true,
+      },
+      styleOverrides: {
+        root: {
+          '&.Mui-focusVisible': {
+            '.MuiSvgIcon-root': {
+              outline: `2px solid ${tokens.borderInputFocus}`,
+              borderRadius: '3px',
+            },
+          },
+          '&.MuiCheckbox-colorPrimary': {
+            color: tokens.colorGrey500,
+            '&:hover': {
+              color: tokens.colorGrey900,
+            },
+            '&.Mui-checked': {
+              color: tokens.colorPrimaryMain,
+              '&:hover': {
+                color: tokens.colorPrimaryDark,
+              },
+            },
+            '&.Mui-disabled': {
+              color: tokens.colorGrey300,
+            },
+          },
+          '&.MuiCheckbox-colorError': {
+            color: tokens.colorGrey500,
+            '&:hover': {
+              color: tokens.colorGrey900,
+            },
+            '&.Mui-checked': {
+              color: tokens.colorErrorMain,
+              '&:hover': {
+                color: tokens.colorErrorDark,
+              },
+            },
+            '&.Mui-disabled': {
+              color: tokens.colorGrey300,
+            },
+          },
+          '.MuiSvgIcon-root': {
+            fontSize: '1.25rem',
+          },
+        },
+      },
+    },
     MuiChip: {
       defaultProps: {
         color: 'default',